requirements.txt
fastapi
uvicorn
openai
pinecone-client
python-dotenv
langchain
